+++ amp/WEB-INF/src/org/digijava/module/dataExchange/utils/ImportBuilder.java
@@ -412,7 +412,7 @@ public class ImportBuilder {
 				
 				AmpActivityProgram ampActivityProgram = new AmpActivityProgram();
 				ampActivityProgram.setActivity(activity);
-				ampActivityProgram.setProgramPercentage((new Float(idmlProgram.getPercentage())).longValue());
+				ampActivityProgram.setProgramPercentage((new Float(idmlProgram.getPercentage())));
 				ampActivityProgram.setProgram(ampTheme);
 				
 				//ampActivityProgram.setProgramSetting()
